+++ sql/CMakeLists.txt
@@ -154,7 +154,7 @@ SET (SQL_SOURCE
                event_queue.cc event_db_repository.cc
                events.cc ../sql-common/my_user.c
                partition_info.cc rpl_utility.cc rpl_utility_server.cc
-               rpl_injector.cc sql_locale.cc
+               rpl_injector.cc sql_locale.cc rpl_info_file.cc
                rpl_rli.cc rpl_mi.cc sql_servers.cc sql_audit.cc
                sql_connect.cc scheduler.cc sql_partition_admin.cc
                sql_profile.cc event_parse_data.cc sql_alter.cc
